+++ tests/Transformers/MaxAbsoluteScalerTest.php
@@ -94,4 +94,17 @@ class MaxAbsoluteScalerTest extends TestCase
         $this->assertNan($samples[0][2]);
         $this->assertNan($samples[1][2]);
     }
+
+    #[Test]
+    public function restoreStateFromSerializedModel() : void
+    {
+        $this->transformer->fit($this->generator->generate(30));
+
+        $this->assertTrue($this->transformer->fitted());
+
+        $restored = unserialize(serialize($this->transformer));
+
+        $this->assertTrue($restored->fitted());
+        $this->assertEquals($this->transformer->maxabs(), $restored->maxabs());
+    }
 }