+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - K Nearest Neighbors and KNN Regressor inference is now parallelized
     - Isolation Forest training and inference is now parallelized
     - Added disk-based neural network snapshotting
+    - Can now clear neural network Adaptive optimizer state
     - Added validation interval parameter to MLPs and GBM Learners
     - Cross Entropy loss function now split into Binary and Multiclass
     - Logistic Regression, Softmax, and Adaline now use hold out set
@@ -34,6 +35,7 @@
     - Removed Word Stemmer tokenizer
     - Added Float Type Converter numeric string and ints to float
     - Boolean Converter now converts truthy and falsy
+    - Interval Discretizer now encodes values as integers
     - Updated to PSR-3 Log version 3
     - Update Amp Backend to Amp version 2.0
     - Added Emoji preset to Regex Filter